refactor: Rename validator package and restructure subpackage in .view
