+++ src/main/java/edu/ntnu/idi/idatt/model/dice/Die.java
@@ -1,6 +1,6 @@
 package edu.ntnu.idi.idatt.model.dice;
 
-import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.dieSetValueValidator;
+import static edu.ntnu.idi.idatt.model.validator.ArgumentValidator.dieSetValueValidator;
 
 import java.util.Random;
