+++ src/main/java/edu/ntnu/idi/idatt/model/tile/SlideAction.java
@@ -2,10 +2,10 @@ package edu.ntnu.idi.idatt.model.tile;
 
 import edu.ntnu.idi.idatt.model.board.Board;
 import edu.ntnu.idi.idatt.model.player.Player;
-import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionPerformValidator;
-import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetDescriptionValidator;
-import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetDestinationTileIdValidator;
-import static edu.ntnu.idi.idatt.model.validators.ArgumentValidator.tileActionSetIdentifierValidator;
+import static edu.ntnu.idi.idatt.model.validator.ArgumentValidator.tileActionPerformValidator;
+import static edu.ntnu.idi.idatt.model.validator.ArgumentValidator.tileActionSetDescriptionValidator;
+import static edu.ntnu.idi.idatt.model.validator.ArgumentValidator.tileActionSetDestinationTileIdValidator;
+import static edu.ntnu.idi.idatt.model.validator.ArgumentValidator.tileActionSetIdentifierValidator;
 /**
  * <h3>SlideAction class</h3>
  *
