+++ src/main/java/edu/ntnu/idi/idatt/navigation/ViewNavigator.java
@@ -13,7 +13,7 @@ import edu.ntnu.idi.idatt.model.board.Board;
 import edu.ntnu.idi.idatt.model.player.Player;
 import edu.ntnu.idi.idatt.observer.ButtonClickObserver;
 import edu.ntnu.idi.idatt.view.app.AppView;
-import edu.ntnu.idi.idatt.view.container.LadderGameBoardCreatorView;
+import edu.ntnu.idi.idatt.view.laddergame.LadderGameBoardCreatorView;
 import edu.ntnu.idi.idatt.view.laddergame.LadderGameView;
 import javafx.scene.Node;
